LPS-96040 Rename
